Switched to GLControl, Added Drag and Drop, new Camera Movement
Switched to OpenTK.GLControl, speeds up rendering, makes movement/zooming less janky.

Added Drag and Drop. Drag one or more files directly onto the form to load them.

Changed up the camera movement.

It now defaults to moving relative to the direction the camera is facing in the X and Z Axis, Kind of like a traditional MMO camera.
Hold Control Key for Flythrough Camera.
Hold Shift key to use the old Axial Movement.

Use "Default to Axis Movement" in the Options menu to set the default movement to Axial Movement, and Shift Key for the new Movement.

The new camera movement can be a little disorenting since the little Axis Marker at it's center doesn't rotate with it.